protocol: ops.DocsAndBytes.docs_total => u64
From u32. Use the same trick we use for bytes_total, so that values
continue to be serialized as native integers rather than using protobuf
string encoding.

+++ crates/proto-flow/src/ops.rs
@@ -191,8 +191,8 @@ pub mod stats {
     #[allow(clippy::derive_partial_eq_without_eq)]
     #[derive(Clone, Copy, PartialEq, ::prost::Message)]
     pub struct DocsAndBytes {
-        #[prost(uint32, tag = "1")]
-        pub docs_total: u32,
+        #[prost(uint64, tag = "1")]
+        pub docs_total: u64,
         #[prost(uint64, tag = "2")]
         pub bytes_total: u64,
     }
